behaviour_acts 자료형의 부동소수점 값을 ".long()" 통해 정수 자료형으로 변경

+++ networks/models.py
@@ -84,7 +84,7 @@ class MlpLSTM(nn.Module):
         batch, seq, *sha = self.batch_size, self.sequence_length, *self.input_size
         hx, cx = lstm_hxs
 
-        behav_acts = behaviour_acts.squeeze(-1)
+        behav_acts = behaviour_acts.long().squeeze(-1)
         
         obs = obs.contiguous().view(batch * seq, *sha)
         x = self.body.forward(obs)
